Fix merge function
The new version merges lights based on their ip address instead of the
ID field. The previous version used the ID, so when the discovery ran
several times, the new generated ID by the "discover" command would
tread the existing light as a new one, and append it to the table.